Update docker-compose.yml with database
I added to my own docker-script the database, so I don't need to mess with postgresql on the bare hardware and starting the container means a fully working installation. Maybe documentation should then add some backup-script for the database (which is actually as easy in docker as for a baremetal db.